move test case summary impl to appropriate file
